requirements.txt
fastapi
uvicorn[standard]
requests
moviepy
ffmpeg-python
python-dotenv
python-multipart


customtkinter

